+++ ndio/utils/parallel.py
@@ -54,24 +54,55 @@ def block_compute(x_start, x_stop,
         Q_stop (int): The upper bound of dimension 'Q'
 
     Returns:
-        :[((x_start, x_stop), (y_start, y_stop), (z_start, z_stop)), ... ]:
+        [((x_start, x_stop), (y_start, y_stop), (z_start, z_stop)), ... ]
     """
-    # First, snap to block in each direction. Then we can divide sans
-    # modulo without having to worry about remainders.
-    bounds = [snap_to_cube(x_start, x_stop, block_size[0], origin[0]),
-              snap_to_cube(y_start, y_stop, block_size[1], origin[1]),
-              snap_to_cube(z_start, z_stop, block_size[2], origin[2])]
+
+    # x
+    x_bounds = range(origin[0], x_stop + block_size[0], block_size[0])
+    x_bounds = [x for x in x_bounds if (x > x_start and x < x_stop)]
+    if len(x_bounds) is 0:
+        x_slices = [(x_start, x_stop)]
+    else:
+        x_slices = []
+        for start_x in x_bounds[:-1]:
+            x_slices.append((start_x, start_x + block_size[0]))
+        x_slices.append((x_start, x_bounds[0]))
+        x_slices.append((x_bounds[-1], x_stop))
+
+    # y
+    y_bounds = range(origin[1], x_stop + block_size[1], block_size[1])
+    y_bounds = [y for y in y_bounds if (y > y_start and y < y_stop)]
+    if len(y_bounds) is 0:
+        y_slices = [(y_start, y_stop)]
+    else:
+        y_slices = []
+        for start_y in y_bounds[:-1]:
+            y_slices.append((start_y, start_y + block_size[1]))
+        y_slices.append((y_start, y_bounds[0]))
+        y_slices.append((y_bounds[-1], y_stop))
+
+    # z
+    z_bounds = range(origin[2], z_stop + block_size[2], block_size[2])
+    z_bounds = [z for z in z_bounds if (z > z_start and z < z_stop)]
+    if len(z_bounds) is 0:
+        z_slices = [(z_start, z_stop)]
+    else:
+        z_slices = []
+        for start_z in z_bounds[:-1]:
+            z_slices.append((start_z, start_z + block_size[2]))
+        z_slices.append((z_start, z_bounds[0]))
+        z_slices.append((z_bounds[-1], z_stop))
+
+    # alright, yuck. but now we have {x, y, z}_slices, each of which hold the
+    # start- and end-points of each cube-aligned boundary. For instance, if you
+    # requested z-slices 4 through 20, it holds [(4, 16), (16, 20)].
+
+    # For my next trick, I'll convert these to a list of:
+    # ((x_start, x_stop), (y_start, y_stop), (z_start, z_stop))
 
     chunks = []
-    for q in range(3):
-        chunks.append([(q_start, q_start + block_size[q])
-                      for q_start in range(bounds[q][0],
-                                           bounds[q][1],
-                                           block_size[q])[:-1]])
-
-    # There's a way to list-comprehensionify the above and the below, but...
-    # a story for a different day.
-    return [(chunks[0][x], chunks[1][y], chunks[2][z])
-            for x in range(len(chunks[0]))
-            for y in range(len(chunks[1]))
-            for z in range(len(chunks[2]))]
+    for x in x_slices:
+        for y in y_slices:
+            for z in z_slices:
+                chunks.append((x, y, z))
+    return chunks